chore: skip corepack test
chore: split corepack e2e tests into two modules

chore: use corepack for pnpm in CI

chore: remove manual install of pnpm

chore: move corepack enable inside install dependencies yml

chore: move corepack install to ci script

chore: use pnpm 8+
