+++ .github/actions/install-node-and-dependencies/action.yml
@@ -28,10 +28,3 @@ runs:
     - name: Install dependencies
       run: ${{ inputs.install-command }}
       shell: bash
-
-    - name: Install pnpm v7
-      run: npm install -g pnpm@7
-      shell: bash
-
-    - run: pnpm --version
-      shell: bash
