+++ .github/workflows/ci.yml
@@ -193,19 +193,19 @@ jobs:
         run: npm ci
         shell: bash
 
+      - name: Enable corepack
+        run: corepack enable
+        shell: bash
+
+      - run: pnpm --version
+        shell: bash
+
       - name: Check if e2e-run-task-runner is affected
         id: check_affected
         run: |
           IS_AFFECTED_STR=$(npx nx print-affected | jq '.projects | contains(["e2e-run-task-runner"])')
           IS_AFFECTED=$([[ $IS_AFFECTED_STR == 'true' ]] && echo 'true' || echo '')
           echo "is_affected=$IS_AFFECTED" >> $GITHUB_OUTPUT
-        shell: bash
-
-      - name: Install pnpm v7
-        run: npm install -g pnpm@7
-        shell: bash
-
-      - run: pnpm --version
         shell: bash
 
       - name: Prepare e2e tests for task-runner
